Move constants to its src folder
